move protocol related types out of shared consts
