+++ packages/protocol-validation/src/types/assets.ts
