+++ packages/protocol-validation/src/types/colors.ts
