+++ packages/protocol-validation/src/types/variables.ts
