+++ packages/shared-consts/src/index.ts
@@ -1,10 +1,4 @@
-export * from "./assets";
-export * from "./codebook";
-export * from "./colors";
 export * from "./controls";
 export * from "./export-process";
 export * from "./network";
-export * from "./protocol";
 export * from "./session";
-export * from "./stages";
-export * from "./variables";
