--- packages/shared-consts/src/protocol.ts
@@ -1,21 +1,0 @@
-import type { Codebook } from "./codebook.js";
-import type { Stage } from "./stages.js";
-
-export type AssetDefinition = {
-	source: string;
-	name: string;
-	type: "network" | "image" | "audio" | "video";
-	id?: string;
-};
-
-export type AssetManifest = Record<string, AssetDefinition>;
-
-export type Protocol = {
-	name: string;
-	description?: string;
-	lastModified: string;
-	schemaVersion: number;
-	stages: Stage[];
-	codebook: Codebook;
-	assetManifest?: AssetManifest;
-};
